+++ backend/src/main/java/middle_point_search/backend/domains/member/domain/RefreshToken.java
@@ -1,4 +1,4 @@
-package middle_point_search.backend.common.redis;
+package middle_point_search.backend.domains.member.domain;
 
 import org.springframework.data.annotation.Id;
 import org.springframework.data.redis.core.RedisHash;
