+++ backend/src/main/java/middle_point_search/backend/domains/member/repository/RefreshTokenRepository.java
@@ -1,9 +1,11 @@
-package middle_point_search.backend.common.redis;
+package middle_point_search.backend.domains.member.repository;
 
 import java.util.Optional;
 
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
+
+import middle_point_search.backend.domains.member.domain.RefreshToken;
 
 @Repository
 public interface RefreshTokenRepository extends CrudRepository<RefreshToken, String> {
